Add a Makefile so test files can run standalone.
Tests may take a while to run sometimes, so this method offers
quicker iteration.
